+++ tests/no_normal_tasks.rs
@@ -60,7 +60,7 @@ fn common_build_configuration(
 ) -> GooseConfiguration {
     if let Some(expect_workers) = manager {
         common::build_configuration(
-            &server,
+            server,
             vec![
                 "--manager",
                 "--expect-workers",
@@ -75,10 +75,10 @@ fn common_build_configuration(
             ],
         )
     } else if worker.is_some() {
-        common::build_configuration(&server, vec!["--worker"])
+        common::build_configuration(server, vec!["--worker"])
     } else {
         common::build_configuration(
-            &server,
+            server,
             vec![
                 "--users",
                 &USERS.to_string(),