+++ tests/scheduler.rs
@@ -120,7 +120,7 @@ fn common_build_configuration(
 ) -> GooseConfiguration {
     if let Some(expect_workers) = manager {
         common::build_configuration(
-            &server,
+            server,
             vec![
                 "--manager",
                 "--expect-workers",
@@ -135,10 +135,10 @@ fn common_build_configuration(
             ],
         )
     } else if worker.is_some() {
-        common::build_configuration(&server, vec!["--worker"])
+        common::build_configuration(server, vec!["--worker"])
     } else {
         common::build_configuration(
-            &server,
+            server,
             vec![
                 "--users",
                 &USERS.to_string(),
@@ -293,7 +293,7 @@ fn run_standalone_test(test_type: &TestType, scheduler: &GooseScheduler) {
     common::run_load_test(goose_attack, None);
 
     // Confirm the load test ran correctly.
-    validate_test(test_type, &scheduler, &mock_endpoints);
+    validate_test(test_type, scheduler, &mock_endpoints);
 }
 
 // Helper to run all gaggle tests.
@@ -373,7 +373,7 @@ fn run_gaggle_test(test_type: &TestType, scheduler: &GooseScheduler) {
     common::run_load_test(manager_goose_attack, Some(worker_handles));
 
     // Confirm the load test ran correctly.
-    validate_test(test_type, &scheduler, &mock_endpoints);
+    validate_test(test_type, scheduler, &mock_endpoints);
 }
 
 #[test]