+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - add test to confirm a `base_url` can include a path and be joined with a relative path
  - fix documentation typo
  - introduce `pretty` log format for `--error-format`, `--debug-format`, `--request-format`, and `--task-format`
+  - clippy cleanups: don't borrow references that are immediately dereferenced by the compiler: https://rust-lang.github.io/rust-clippy/master/index.html#needless_borrow
 
 ## 0.13.0 July 19, 2021
   - enable [`gzip`](https://docs.rs/reqwest/*/reqwest/struct.ClientBuilder.html#method.gzip) support and set Accept-Encoding header by default in the client; disable with `--no-gzip` or `GooseDefault::NoGzip`